requirements.txt
fastapi
uvicorn[standard]
presidio-analyzer
presidio-anonymizer
spacy
pyyaml
httpx